Version 3.0.0
Not much on in the front end, but a complete revamp in the back.
Collection system completely scrapped and replaced with new (was long overdue).
Notifications were also revamped in the backend but that needs improvement but a defining base has been set.
Stats were introduced a couple of version back but were again scrapped for something alot better. Using JS to track impressions and clicks.
Huge update so might be alot of bugs, will try to check for them update.